rag-service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
chromadb==1.0.20
llama-index-core==0.13.2
llama-index-llms-openai==0.5.6
llama-index-embeddings-openai==0.5.1
llama-index-vector-stores-chroma==0.5.0
pypdf==6.0.0
python-docx==1.2.0
